Make tools command compliant with PSR-4 autoloading
