+++ tools/src/Command/CheckDatabaseKeysCommand.php
@@ -33,13 +33,15 @@
  * ---------------------------------------------------------------------
  */
 
+namespace Glpi\Tools\Command;
+
 use Glpi\Console\AbstractCommand;
 use Glpi\System\Diagnostic\DatabaseKeysChecker;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
-class CheckDatabaseKeysCommand extends AbstractCommand
+final class CheckDatabaseKeysCommand extends AbstractCommand
 {
     /**
      * Error code returned when missing keys are found.
