+++ tools/src/Command/CheckTwigTemplatesSyntaxCommand.php
@@ -45,7 +45,7 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Twig\Error\Error;
 
-class CheckTwigTemplatesSyntaxCommand extends Command
+final class CheckTwigTemplatesSyntaxCommand extends Command
 {
     /**
      * Error code returned when some templates have invalid yntax.
